% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairing.R
\name{cantor_pairing}
\alias{cantor_pairing}
\title{Uniquely encode two non-negative integers to a single non-negative integer.}
\usage{
cantor_pairing(x, y)
}
\arguments{
\item{x}{A non-negative integer.}

\item{y}{A non-negative integer.}
}
\value{
A non-negative integer.
}
\description{
This function uniquely encodes two non-negative integers to a single
non-negative integer, using the Cantor pairing function. See the \href{https://en.wikipedia.org/wiki/Pairing_function#Cantor_pairing_function}{Wikipedia article}
for more information.
}
\examples{
cantor_pairing(0, 0)
cantor_pairing(47, 32)
}

