% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairing.R
\name{hu_pairing}
\alias{hu_pairing}
\title{Uniquely encode two positive integers to a single non-negative integer.}
\usage{
hu_pairing(x, y)
}
\arguments{
\item{x}{A positive integer.}

\item{y}{A positive integer.}
}
\value{
A positive integer.
}
\description{
This function uniquely encodes two positive integers to a single non-negative
integer, using the Hopcrof-Ullman pairing function.
See the \href{http://mathworld.wolfram.com/PairingFunction.html}{MathWorld article}
for more information.
}
\examples{
hu_pairing(1, 1)
}

