% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairing.R
\name{inverse_cantor_pairing}
\alias{inverse_cantor_pairing}
\title{Invert the Cantor pairing function.}
\usage{
inverse_cantor_pairing(z)
}
\arguments{
\item{z}{A non-negative integer}
}
\value{
A vector of non-negative integers \code{(x, y)} such that
   \code{cantor_pairing(x, y) == z}.
}
\description{
This function is the inverse to the Cantor pairing function. See the \href{https://en.wikipedia.org/wiki/Pairing_function#Cantor_pairing_function}{Wikipedia article}
for more information.
}
\examples{
inverse_cantor_pairing(0)
inverse_cantor_pairing(3192)
}

