% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairing.R
\name{inverse_hu_pairing}
\alias{inverse_hu_pairing}
\title{Invert the Hopcroft-Ullman pairing function.}
\usage{
inverse_hu_pairing(z)
}
\arguments{
\item{z}{A non-negative integer}
}
\value{
A vector of non-negative integers \code{(x, y)} such that
   \code{cantor_pairing(x, y) == z}.
}
\description{
This function is the inverse to the Hopcroft-Ullman pairing function.
See the \href{http://mathworld.wolfram.com/PairingFunction.html}{MathWorld article}
for more information.
}
\examples{
inverse_hu_pairing(1)
}

